Split the agentic tool into translate_agentic and verify_fix_agentic
