+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["benchmark", "core", "tools/full_source", "tools/build_project_spec", "tools/load_raw_source", "tools/raw_source_to_cargo_llm", "tools/try_cargo_build", "translate", "tools/c_ast", "tools/modular_translation_llm", "tools/quantize_rust_spans" , "tools/fix_declarations_llm", "tools/agentic"]
+members = ["benchmark", "core", "tools/full_source", "tools/build_project_spec", "tools/load_raw_source", "tools/raw_source_to_cargo_llm", "tools/try_cargo_build", "translate", "tools/c_ast", "tools/modular_translation_llm", "tools/quantize_rust_spans", "tools/fix_declarations_llm", "tools/translate_agentic", "tools/verify_fix_agentic"]
 resolver = "3"
 
 [workspace.dependencies]
@@ -14,7 +14,8 @@ try_cargo_build = { path = "tools/try_cargo_build" }
 c_ast = { path = "tools/c_ast" }
 modular_translation_llm = { path = "tools/modular_translation_llm" }
 quantize_rust_spans = { path = "tools/quantize_rust_spans" }
-agentic = { path = "tools/agentic" }
+translate_agentic = { path = "tools/translate_agentic" }
+verify_fix_agentic = { path = "tools/verify_fix_agentic" }
 log = "0.4.28"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
