+++ benchmark/src/main.rs
@@ -74,6 +74,7 @@ pub fn translate_c_directory_to_rust_project(
     config_overrides: &[String],
     modular: bool,
     agentic: bool,
+    agentic_verify: bool,
 ) -> TranspilationResult {
     let args: Arc<harvest_translate::cli::Args> = harvest_translate::cli::Args {
         input: Some(input_dir.to_path_buf()),
@@ -83,6 +84,7 @@ pub fn translate_c_directory_to_rust_project(
         force: false,
         modular,
         agentic,
+        agentic_verify,
     }
     .into();
     let mut config = harvest_translate::cli::initialize(args).expect("Failed to generate config");
@@ -131,6 +133,7 @@ pub fn run_all_benchmarks(
     timeout: u64,
     modular: bool,
     agentic: bool,
+    agentic_verify: bool,
 ) -> HarvestResult<Vec<ProgramEvalStats>> {
     // Process all examples
     let mut results = Vec::new();
@@ -142,7 +145,7 @@ pub fn run_all_benchmarks(
         log::info!("{}", "=".repeat(80));
 
         let result =
-            benchmark_single_program(program_dir, output_dir, config_overrides, timeout, modular, agentic);
+            benchmark_single_program(program_dir, output_dir, config_overrides, timeout, modular, agentic, agentic_verify);
 
         results.push(result);
     }
@@ -227,6 +230,7 @@ fn benchmark_single_program(
     timeout: u64,
     modular: bool,
     agentic: bool,
+    agentic_verify: bool,
 ) -> ProgramEvalStats {
     let program_name = program_dir
         .file_name()
@@ -281,6 +285,7 @@ fn benchmark_single_program(
         config_overrides,
         modular,
         agentic,
+        agentic_verify,
     );
 
     result.translation_success = translation_result.translation_success;
@@ -472,6 +477,7 @@ fn run(args: Args) -> HarvestResult<()> {
         args.timeout,
         args.modular,
         args.agentic,
+        args.agentic_verify,
     )?;
     let csv_output_path = args.output_dir.join("results.csv");
     write_csv_results(&csv_output_path, &results)?;
