--- tools/agentic/src/translate_single.rs
@@ -1,134 +1,0 @@
-//! Single-case translation via kiro-cli.
-//!
-//! Translates one C project (presented as a directory of source files) into a Rust Cargo project
-//! by invoking `kiro-cli` as an external agent. After the agent finishes, the generated
-//! `Cargo.toml` is post-processed to satisfy downstream tool expectations.
-
-use crate::cargo_toml::{self, CargoToml};
-
-use build_project_spec::ProjectKind;
-use std::path::Path;
-use std::process::Command;
-use tracing::{info, warn};
-
-type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
-
-/// Default kiro-cli translation timeout in seconds (30 minutes).
-const TRANSLATE_TIMEOUT_SECS: u64 = 1800;
-
-/// Runs the translation agent on a single C project.
-///
-/// # Directory layout
-///
-/// Before this function is called, the caller must set up:
-///
-/// ```text
-/// case_dir/
-///   translated_rust/
-///     c_src/          <- C source files (materialized RawSource)
-/// ```
-///
-/// After a successful return, the directory will additionally contain:
-///
-/// ```text
-/// case_dir/
-///   translated_rust/
-///     Cargo.toml      <- generated by kiro-cli, post-processed
-///     src/            <- generated by kiro-cli
-///     c_src/          <- unchanged
-///   translated_rust_original/  <- snapshot for verify stage
-/// ```
-pub fn translate(
-    case_dir: &Path,
-    prompt: &str,
-    project_kind: &ProjectKind,
-) -> Result<()> {
-    let translated = case_dir.join("translated_rust");
-
-    invoke_kiro_cli(&translated, prompt, TRANSLATE_TIMEOUT_SECS)?;
-
-    if !translated.join("Cargo.toml").exists() {
-        return Err("kiro-cli did not produce a Cargo.toml".into());
-    }
-
-    post_process(&translated, project_kind)?;
-    save_original(case_dir)?;
-
-    info!("Translation complete");
-    Ok(())
-}
-
-/// Invokes kiro-cli in `work_dir` with the given prompt and timeout.
-fn invoke_kiro_cli(work_dir: &Path, prompt: &str, timeout_secs: u64) -> Result<()> {
-    info!("Invoking kiro-cli (timeout={}s)", timeout_secs);
-
-    let logs_dir = work_dir.parent().unwrap_or(work_dir).join("logs");
-    std::fs::create_dir_all(&logs_dir)?;
-    let log_path = logs_dir.join("translation.log");
-
-    let status = Command::new("bash")
-        .arg("-c")
-        .arg(format!(
-            "set -o pipefail; timeout {timeout_secs} kiro-cli chat \
-             --no-interactive --trust-all-tools \"$PROMPT\" < /dev/null 2>&1 | tee \"$LOG\"",
-        ))
-        .env("PROMPT", prompt)
-        .env("LOG", &log_path)
-        .env(
-            "OPENSSL_DIR",
-            std::env::var("OPENSSL_DIR").unwrap_or_else(|_| "/usr".into()),
-        )
-        .current_dir(work_dir)
-        .status()?;
-
-    if !status.success() {
-        warn!("kiro-cli exited with {status}");
-    }
-    Ok(())
-}
-
-/// Applies standard Cargo.toml fixups after the agent finishes.
-fn post_process(translated: &Path, project_kind: &ProjectKind) -> Result<()> {
-    let cargo_path = translated.join("Cargo.toml");
-    let mut cargo = CargoToml::open(&cargo_path)?;
-
-    cargo.add_workspace();
-
-    match project_kind {
-        ProjectKind::Library => {
-            cargo.remove_bin();
-            // Use the package name already chosen by the agent as the library name.
-            if let Some(name) = read_package_name(&cargo_path) {
-                cargo.set_lib(&name);
-            }
-            cargo.save()?;
-            cargo_toml::strip_for_lib(translated)?;
-        }
-        ProjectKind::Executable => {
-            cargo.set_bin_driver();
-            cargo.save()?;
-        }
-    }
-    Ok(())
-}
-
-/// Reads `[package].name` from a Cargo.toml. Returns `None` on any failure.
-fn read_package_name(manifest: &Path) -> Option<String> {
-    let contents = std::fs::read_to_string(manifest).ok()?;
-    let doc: toml_edit::DocumentMut = contents.parse().ok()?;
-    doc.get("package")?
-        .get("name")?
-        .as_str()
-        .map(|s| s.to_string())
-}
-
-/// Copies `translated_rust/` to `translated_rust_original/` as a snapshot for the verify stage.
-fn save_original(case_dir: &Path) -> Result<()> {
-    let translated = case_dir.join("translated_rust");
-    let original = case_dir.join("translated_rust_original");
-    if original.exists() {
-        std::fs::remove_dir_all(&original)?;
-    }
-    harvest_core::cargo_utils::copy_directory_recursive(&translated, &original)?;
-    Ok(())
-}
