+++ tools/translate_agentic/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "translate_agentic"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+build_project_spec.workspace = true
+full_source.workspace = true
+harvest_core.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tempfile.workspace = true
+toml_edit.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
