+++ tools/translate_agentic/src/cargo_toml.rs
