+++ tools/translate_agentic/src/prompt_executable.md
