+++ tools/verify_fix_agentic/Cargo.toml
@@ -1,16 +1,14 @@
 [package]
-name = "agentic"
+name = "verify_fix_agentic"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-build_project_spec.workspace = true
 full_source.workspace = true
 harvest_core.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tempfile.workspace = true
-toml_edit.workspace = true
 tracing.workspace = true
 
 [lints]
