+++ translate/Cargo.toml
@@ -27,7 +27,8 @@ build_project_spec.workspace = true
 modular_translation_llm.workspace = true
 try_cargo_build.workspace = true
 raw_source_to_cargo_llm.workspace = true
-agentic.workspace = true
+translate_agentic.workspace = true
+verify_fix_agentic.workspace = true
 
 [lints]
 workspace = true
