+++ translate/src/lib.rs
@@ -6,7 +6,6 @@ mod runner;
 mod scheduler;
 pub mod util;
 
-use agentic::AgenticTool;
 use build_project_spec::BuildProjectSpec;
 use c_ast::ParseToAst;
 use harvest_core::config::Config;
@@ -19,7 +18,9 @@ use runner::ToolRunner;
 use scheduler::Scheduler;
 use std::sync::Arc;
 use tracing::{error, info};
+use translate_agentic::TranslateAgentic;
 use try_cargo_build::TryCargoBuild;
+use verify_fix_agentic::VerifyFixAgentic;
 
 /// Performs the complete transpilation process using the scheduler.
 pub fn transpile(config: Arc<Config>) -> Result<HarvestIR, Box<dyn std::error::Error>> {
@@ -36,7 +37,12 @@ pub fn transpile(config: Arc<Config>) -> Result<HarvestIR, Box<dyn std::error::E
     let load_src = scheduler.queue(LoadRawSource::new(&config.input));
     let project_spec = scheduler.queue_after(BuildProjectSpec, &[load_src]);
     let translate = if config.agentic {
-        scheduler.queue_after(AgenticTool, &[load_src, project_spec])
+        let t = scheduler.queue_after(TranslateAgentic, &[load_src, project_spec]);
+        if config.agentic_verify {
+            scheduler.queue_after(VerifyFixAgentic, &[t, load_src])
+        } else {
+            t
+        }
     } else if config.modular {
         let parse_ast = scheduler.queue_after(ParseToAst, &[load_src]);
         scheduler.queue_after(ModularTranslationLlm, &[load_src, parse_ast, project_spec])
